+++ cmd/network_validator_deposit_alt.go
@@ -21,22 +21,19 @@ address. Remember it will need your wallet address and private keys. Thus it wil
 This tool is necessary to activate new validators`,
 	Example: "lukso network validator deposit_alt",
 	Run: func(cmd *cobra.Command, args []string) {
-		nodeConf, err := network.GetLoadedNodeConfigs()
-		if err != nil {
-			cobra.CompErrorln(err.Error())
-			return
-		}
+		nodeConf := network.MustGetNodeConfig()
 		valSecrets := nodeConf.GetValSecrets()
 		if valSecrets == nil {
 			cobra.CompErrorln("no validator credential is presented")
 			return
 		}
 
-		if err := network.Deposit(valSecrets.Deposit.DepositFileLocation, valSecrets.Deposit.ContractAddress, valSecrets.Eth1Data.WalletPrivKey, nodeConf.ApiEndpoints.ExecutionApi); err != nil {
+		totalDeposits, err := network.Deposit(valSecrets.Deposit.DepositFileLocation, valSecrets.Deposit.ContractAddress, valSecrets.Eth1Data.WalletPrivKey, nodeConf.ApiEndpoints.ExecutionApi)
+		if err != nil {
 			cobra.CompErrorln(err.Error())
 			return
 		}
-		fmt.Println("You successfully deposited your key(s)! Your keys need to be activated which takes around 8h. You can check the status by calling:")
+		fmt.Printf("You successfully deposited %d key(s)! Your keys need to be activated which takes around 8h. You can check the status by calling:\n", totalDeposits)
 		fmt.Println(utils.ConsoleInBlue("        lukso network validator describe"))
 	},
 }