+++ src/network/node_config.go
@@ -1,6 +1,9 @@
 package network
 
 import (
+	"fmt"
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 	"os"
 
 	"gopkg.in/yaml.v3"
@@ -158,5 +161,77 @@ func (nc *NodeConfigs) WriteOrUpdateNodeConfig() error {
 	if err != nil {
 		return err
 	}
-	return os.WriteFile("./node_config.yaml", yamlData, os.ModePerm)
+	return os.WriteFile(NodeConfigLocation, yamlData, os.ModePerm)
+}
+
+func (nc *NodeConfigs) UpdateBootnodes() error {
+	chain := GetChainByString(nc.Chain.Name)
+	GetChainByString(nc.Chain.Name)
+	bootnodes, err := NewBootnodeUpdater(chain).DownloadLatestBootnodes()
+	if err != nil {
+		return err
+	}
+
+	if len(bootnodes) == 0 {
+		fmt.Println("No bootnodes available for this chain ", chain.String())
+	}
+
+	hasUpdates := false
+	if nc.Consensus.Bootnode != bootnodes[0].Consensus {
+		fmt.Println("Updating bootnode for the consensus chain...")
+		hasUpdates = true
+		nc.Consensus.Bootnode = bootnodes[0].Consensus
+	}
+	if nc.Execution.Bootnode != bootnodes[0].Execution {
+		fmt.Println("Updating bootnode for the execution chain...")
+		hasUpdates = true
+		nc.Execution.Bootnode = bootnodes[0].Execution
+	}
+
+	if !hasUpdates {
+		fmt.Println("everything up to date")
+	} else {
+		err := nc.WriteOrUpdateNodeConfig()
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func MustGetNodeConfig() *NodeConfigs {
+	if !FileExists(NodeConfigLocation) {
+		cobra.CompErrorln("The node was not initialised yet. Call \n   lukso network init --nodeName NAME_OF_YOUR_NODE [--chain CHAIN_NAME] \n to setup a node.")
+		os.Exit(1)
+	}
+
+	// Search config in home directory with name ".cli" (without extension).
+	viper.AddConfigPath(".")
+	viper.SetConfigType("yaml")
+	viper.SetConfigName("node_config")
+
+	viper.AutomaticEnv()
+
+	if err := viper.ReadInConfig(); err == nil {
+		fmt.Println("config file:", viper.ConfigFileUsed())
+	} else {
+		cobra.CompErrorln(err.Error())
+		os.Exit(1)
+	}
+
+	config, err := getLoadedNodeConfigs()
+	if err != nil {
+		cobra.CompErrorln(err.Error())
+		os.Exit(1)
+	}
+	return config
+}
+
+func getLoadedNodeConfigs() (*NodeConfigs, error) {
+	var nodeConfig NodeConfigs
+	err := viper.Unmarshal(&nodeConfig)
+	if err != nil {
+		return nil, err
+	}
+	return &nodeConfig, nil
 }